Fix bug where incorrect class was being searched for static factory methods.
The class declared as the type of an instance was being searched for the
factory method.  This caused problems when the factory method for a bean
was not in the same class as the declared type of the bean.
This change searches the 'referenced type' for the factory method instead.