Add ConfigurationBuilder to wire SpringConfigurationLoader and augmentations.
This adds a ConfigurationBuilder class to simplify creating a
SpringConfigurationLoader configured with the necessary augmentations.
The ConfigurationBuilder supports ApplicationContext objects in
addition to BeanFactory objects.

This change also adds the BeanFactory as an instance variable of
SpringConfigurationLoader. The BeanFactory is required for retrieving
information about each bean. Previously, it was only available for
top-level beans. By storing the BeanFactory as an instance variable,
it is now available for obtaining the BeanDefinition of nested beans.