Merge nested bean definitions with their parent bean definitions.
Previously, only top-level beans were merged with parent definitions.
The Spring getMergedBeanDefinition method cannot be used for bean
definitions that are not at the top-level, because the BeanFactory does
not recognise the name of the nested bean.

This change includes a simple algorithm for merging nested beans with
their parents, because the methods required to merge BeanDefinitions
in the BeanFactory are protected methods.